Fixed #513, translate tutorial.
