+++ package.json
@@ -22,7 +22,8 @@
         "tutorial-schema": "npx ts-json-schema-generator --no-type-check --tsconfig tsconfig.json --path 'src/tutorial/Tutorial.ts' --type 'Tutorial' --out static/schemas/Tutorial.json",
         "locale-schema": "npx ts-json-schema-generator --no-type-check --tsconfig tsconfig.json --path 'src/locale/LocaleText.ts' --type 'LocaleText' --out static/schemas/LocaleText.json",
         "create-schemas": "npm run tutorial-schema && npm run locale-schema",
-        "verify-locales": "npx tsx src/util/verify-locales.ts",
+        "verify-locales": "npx tsx src/util/verify-locales/start.ts verify",
+        "translate-locales": "npx tsx src/util/verify-locales/start.ts translate",
         "locales": "npm-watch verify-locales",
         "schemas": "npm-watch create-schemas",
         "merge": "git checkout main && git merge dev && git push origin main && git checkout dev"
