+++ src/util/verify-locales/Validator.ts
@@ -1,0 +1,6 @@
+import Ajv from 'ajv';
+
+// Initialize the schema validator.
+const Validator = new Ajv({ strictTuples: false, allErrors: true });
+
+export default Validator;
